@@ -13,11 +13,16 @@ namespace Oveleon\ContaoComponentStyleManager\StyleManager;
 use Contao\StringUtil;
 use Contao\System;
 use Oveleon\ContaoComponentStyleManager\Controller\BackendModule\ImportController;
+use Oveleon\ContaoComponentStyleManager\Model\StyleManagerArchiveModel;
+use Oveleon\ContaoComponentStyleManager\Model\StyleManagerModel;
+use Symfony\Component\Yaml\Yaml;
 
 /**
  * A static class to store config data
  *
  * @author Daniele Sciannimanica <https://github.com/doishub>
+ *
+ * @internal
  */
 final class Config
 {
@@ -105,11 +110,13 @@ final class Config
     /**
      * Return all configuration files from third-party bundles
      */
-    public static function getBundleConfigurationFiles(): array|null
+    public static function getBundleConfigurationFiles(ConfigurationFileType $type): array|null
     {
+        $container = System::getContainer();
+
         /** @var string $projectDir */
-        $projectDir = System::getContainer()->getParameter('kernel.project_dir');
-        $arrFiles = System::getContainer()->get('contao.resource_finder')?->findIn('templates')?->files()?->name('style-manager-*.xml');
+        $projectDir = $container->getParameter('kernel.project_dir');
+        $arrFiles = $container->get('contao.resource_finder')?->findIn('templates')?->files()?->name('style-manager-*'. $type->value);
         $arrBundleConfigs = null;
 
         if ($arrFiles->hasResults())
@@ -161,13 +168,173 @@ final class Config
      */
     protected function loadBundleConfiguration(): array
     {
+        $configuration = [[],[]];
+
         if (
-            ($arrFiles = $this->getBundleConfigurationFiles())
-            && (is_array($bundleConfig = ImportController::importFiles($arrFiles, false)))
+            ($xmlFiles = $this->getBundleConfigurationFiles(ConfigurationFileType::XML))
+            && (is_array($xmlBundleConfig = ImportController::importXmlFiles($xmlFiles, false)))
         ) {
-            return $bundleConfig;
+            $configuration = $xmlBundleConfig;
         }
 
-        return [[],[]];
+        if ($yamlFiles = $this->getBundleConfigurationFiles(ConfigurationFileType::YAML))
+        {
+            $configuration = self::parseYamlConfiguration($yamlFiles, $configuration);
+        }
+
+        return $configuration;
+    }
+
+    /**
+     * @experimental
+     */
+    private function parseYamlConfiguration(array $files, array $configuration): array
+    {
+        [$styleArchives, $styleGroups] = $configuration;
+
+        foreach ($files as $filePath)
+        {
+            if (
+                !is_file($filePath)
+                || !is_readable($filePath)
+                || null === ($archives = Yaml::parseFile($filePath))
+            ) {
+                continue;
+            }
+
+            foreach ($archives as $archiveIdent => $data)
+            {
+                if (isset($styleArchives[$archiveIdent]))
+                {
+                    $archive = $styleArchives[$archiveIdent];
+                }
+                else
+                {
+                    $archive = new StyleManagerArchiveModel();
+                    $archive->identifier = $archiveIdent;
+                }
+
+                foreach ($data as $k => $v)
+                {
+                    if ($k === 'children')
+                    {
+                        continue;
+                    }
+
+                    $archive->{$k} = $v;
+                }
+
+                $styleArchives[$archiveIdent] = $archive;
+
+                if (!isset($data['children']))
+                {
+                    continue;
+                }
+
+                foreach ($data['children'] as $childAlias => $childData)
+                {
+                    $styleIdent = StyleManager::generateAlias($archiveIdent, $childAlias);
+
+                    if (isset($styleGroups[$styleIdent]))
+                    {
+                        $child = $styleGroups[$styleIdent];
+                    }
+                    else
+                    {
+                        $child = new StyleManagerModel();
+                        $child->alias = $childAlias;
+
+                        // Fake the pid for the ComponentStyleSelect Widget
+                        $child->pid = $archiveIdent;
+                    }
+
+                    // ToDo: Generalize that in future with the part in the ImportController
+                    foreach ($childData as $kk => $vv)
+                    {
+                        if ($kk === 'id' || !$vv || (!is_array($vv) && strtolower((string) $vv) === 'null'))
+                        {
+                            continue;
+                        }
+
+                        switch ($kk)
+                        {
+                            case 'pid':
+                                $vv = $child->pid;
+                                break;
+
+                            case 'cssClasses':
+                                if (is_array($vv))
+                                {
+                                    $vv = $this->convertCssClasses($vv);
+                                }
+
+                                if ($child->{$kk})
+                                {
+                                    /** @var array<array<int|string>> $arrClasses */
+                                    $arrClasses = StringUtil::deserialize($child->{$kk}, true);
+                                    $arrExists  = Sync::flattenKeyValueArray($arrClasses);
+
+                                    /** @var array<array<int|string>> $arrValues */
+                                    $arrValues  = StringUtil::deserialize($vv, true);
+
+                                    foreach ($arrValues as $cssClass)
+                                    {
+                                        if (!array_key_exists($cssClass['key'], $arrExists))
+                                        {
+                                            $arrClasses[] = [
+                                                'key'   => $cssClass['key'],
+                                                'value' => $cssClass['value']
+                                            ];
+                                        }
+                                    }
+
+                                    $vv = serialize($arrClasses);
+                                }
+                                break;
+
+                            default:
+                                $dcaField = $GLOBALS['TL_DCA']['tl_style_manager']['fields'][$kk] ?? null;
+
+                                if (isset($dcaField['eval']['multiple']) && !!$dcaField['eval']['multiple'] && $dcaField['inputType'] === 'checkbox')
+                                {
+                                    /** @var array<array<int|string>> $arrElements */
+                                    $arrElements = StringUtil::deserialize($child->{$kk}, true);
+
+                                    /** @var array<array<int|string>> $arrValues */
+                                    $arrValues = StringUtil::deserialize($vv, true);
+
+                                    foreach ($arrValues as $element)
+                                    {
+                                        if (!in_array($element, $arrElements, true))
+                                        {
+                                            $arrElements[] = $element;
+                                        }
+                                    }
+
+                                    $vv = serialize($arrElements);
+                                }
+                        }
+
+                        $child->{$kk} = $vv;
+                    }
+
+                    $styleGroups[$styleIdent] = $child;
+                }
+            }
+        }
+
+        return [$styleArchives, $styleGroups];
+    }
+
+    private function convertCssClasses(array $array): array
+    {
+        $return = [];
+
+        foreach ($array as $class => $label)
+        {
+            $return[] = ['key' => $class, 'value' => $label];
+        }
+
+        return $return;
     }
 }
